Auth almost done!
made a huge progress toward making the auth work completly
the sign up works, the sign in doesn't for now. learned a lot about the importance of modules especially prisma.module , from now on its better to make a resource for every thing to improve the quality of the code.